% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellOnTools-package.R
\docType{package}
\name{CellOnTools-package}
\alias{CellOnTools}
\alias{CellOnTools-package}
\title{CellOnTools: Tools for Cell Ontology Analysis}
\description{
CellOnTools provides utilities for mapping, querying, visualising, and
aggregating Cell Ontology (CL) terms, as well as marker-gene enrichment using
bundled healthy human and mouse CellMarkerAccordion annotations.
}
\details{
Main functionality includes:
\itemize{
\item Loading the fixed 2026-06-08 Cell Ontology release, or an explicitly
selected AnnotationHub/local OBO source.
\item Converting between CL identifiers and labels.
\item Searching and mapping free-text cell type names to CL terms.
\item Extracting and plotting ontology hierarchy subgraphs.
\item Computing semantic similarity between CL terms.
\item Rolling fine-grained CL terms up to coarser shared ancestors.
\item Performing marker-gene enrichment against bundled marker annotations.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WANGDI0212/CellOnTools}
  \item Report bugs at \url{https://github.com/WANGDI0212/CellOnTools/issues}
}

}
\author{
\strong{Maintainer}: Di Wang \email{wangdi0001@foxmail.com}

}
\keyword{internal}
